Add builders for newer major versions of ubuntu
Why do we need this change?
=======================
This change facilitates the ability for folks to run nominatim on a
version of ubuntu that they wish to. Additionally this sets up the
ability for folks to run external databases that match the version by
copying the nominatim.so library from these containers and running on
the appropriate database version

What effects does this change have?
=======================
* Moves from single Docker file to a postgresql directory based approach
* Add ubuntu bionic (postgresql 10) support
* Add ubuntu focal (postgres 12) support
* create a common_copy.sh function that allows updating all individual
  builds to be updated when local.php/nominatim.conf are modified